Refactored the process scheduling component of Input handler into it's own class.
